Strange attendance system
Raghav lives in Strangeland and studies in a college where it is mandatory to maintain a minimum of 75% attendance in each subject or else you are not allowed to appear for the exam of that subject. Consider the subject AI (Artificial Intelligence) for which the total number of classes held till the present date equals . ​Raghav has only attended  out of these  classes. Find minimum how many extra classes should be held so that he is allowed to appear for the AI exam. Assume that he attends all the extra classes. 

INPUT FORMAT

First line consists of a single integer  denoting the number of testcases

Next  lines each consist of two space separated integers  and  as described above
OUTPUT FORMAT

For each testcase output a single integer denoting the answer for that testcase

CONSTRAINTS



SUBTASKS

For 20 Points : 
For 80 Points : Original Constraints
SAMPLE INPUT 
5
7 6
9 1
6 3
10 5
2 1
SAMPLE OUTPUT 
0
23
6
10
2
Explanation
For the first testcase : 7 classes have been held out of which Raghav has attended 6 which is greater than 75% .Hence no more extra classes are required.